+++ CHANGELOG.md
@@ -14,6 +14,10 @@
 - Add joinSource field for joinGroup method
 - Add advanced revoke type
 - Add recvMessageRevokedV2 callback for OnAdvancedMsgListener
+- Add joinSource field for GroupApplicationInfo
+- Add inviterUserID field for GroupApplicationInfo
+- Add getAdvancedHistoryMessageList method for MessageManager
+- Add searchGroupMembers method for GroupManager
 
 ## 2.2.0